Extract ApplicationStartupManager from AppDelegate
